# server/openapi.yaml
openapi: 3.1.0
info:
  title: 题库系统
  description: ""
  version: 1.0.0
servers:
  - url: http://127.0.0.1:6020
    description: 开发环境
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Auditor:
      type: object
      properties:
        created_at:
          type: string
          format: date-time
          description: 创建时间
        created_by:
          type: integer
          format: int64
          description: 创建人
        updated_at:
          type: string
          format: date-time
          description: 更新时间
        updated_by:
          type: integer
          format: int64
          description: 更新人
      required:
        - created_at
        - created_by
        - updated_at
        - updated_by
    ID:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID
      required:
        - id
    IDS:
      type: array
      items:
        type: integer
        format: int64
...
    MenuType:
      type: string
      enum:
        - M
        - C
        - F
      x-enum-descriptions:
        - 目录
        - 菜单
        - 按钮
      x-enum-varnames:
        - directory
        - menu
        - button
...
    ApiCreate:
      type: object
      properties:
        handle:
          type: string
          description: 接口标识
        title:
          type: string
          description: 接口名称
        path:
          type: string
          description: 接口路径
        type:
          type: string
          description: 接口类型
        action:
          type: string
          description: 接口动作
      required:
        - handle
        - title
        - path
        - type
        - action
...
    Api:
      allOf:
        - $ref: "#/components/schemas/ID"
        - $ref: "#/components/schemas/ApiCreate"
        - $ref: "#/components/schemas/Auditor"
...
    Tree:
      type: object
      properties:
        id:
          type: integer
          format: int64
        label:
          type: string
        children:
          type: array
          items:
            $ref: "#/components/schemas/Tree"
      required:
        - id
        - label
        - children
...
    BaseMenu:
      type: object
      properties:
        title:
          type: string
          description: 菜单标题
        parent_id:
          type: integer
          format: int64
          description: 父级菜单ID
        type:
          $ref: "#/components/schemas/MenuType"
        sort:
          type: string
          description: 菜单排序
      required:
        - title
        - parent_id
        - type
    MenuWithDirectory:
      allOf:
        - $ref: "#/components/schemas/BaseMenu"
      properties:
        name:
          type: string
          description: 路由名称
        icon:
          type: string
          description: 菜单图标
        route:
          type: string
          description: 路由路径
        visible:
          type: boolean
          description: 是否可见
      required:
        - name
        - route
    MenuWithMenu:
      allOf:
        - $ref: "#/components/schemas/MenuWithDirectory"
      properties:
        permission:
          type: string
          description: 权限标识
        api_ids:
          $ref: "#/components/schemas/IDS"
    MenuWithButton:
      allOf:
        - $ref: "#/components/schemas/BaseMenu"
      properties:
        permission:
          type: string
          description: 权限标识
        api_ids:
          $ref: "#/components/schemas/IDS"
    MenuCreate:
      oneOf:
        - $ref: "#/components/schemas/MenuWithDirectory"
        - $ref: "#/components/schemas/MenuWithMenu"
        - $ref: "#/components/schemas/MenuWithButton"
      discriminator:
        propertyName: type
        mapping:
          "M": "#/components/schemas/MenuWithDirectory"
          "C": "#/components/schemas/MenuWithMenu"
          "F": "#/components/schemas/MenuWithButton"
    MenuUpdate:
      oneOf:
        - $ref: "#/components/schemas/MenuWithDirectory"
        - $ref: "#/components/schemas/MenuWithMenu"
        - $ref: "#/components/schemas/MenuWithButton"
      discriminator:
        propertyName: type
        mapping:
          "M": "#/components/schemas/MenuWithDirectory"
          "C": "#/components/schemas/MenuWithMenu"
          "F": "#/components/schemas/MenuWithButton"
    Menu:
      allOf:
        - $ref: "#/components/schemas/ID"
        - $ref: "#/components/schemas/MenuCreate"
        - $ref: "#/components/schemas/Auditor"
      properties:
        apis:
          type: array
          items:
            $ref: "#/components/schemas/Api"
        children:
          type: array
          items:
            $ref: "#/components/schemas/Menu"
      required:
        - children
        - sort
    MenuCompact:
      type: object
      properties:
        id:
          type: integer
          format: int64
        label:
          type: string
        icon:
          type: string
        route:
          type: string
          description: 路由路径
        children:
          type: array
          items:
            $ref: "#/components/schemas/MenuCompact"
      required:
        - id
        - label
        - route
        - children
...
security:
  - bearerAuth: [ ]
...
paths:
...
  /menus:
    post:
      summary: 创建菜单
      operationId: createMenu
      tags:
        - menu
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuCreate"
      responses:
        "201":
          description: "创建成功"
      security:
        - bearer: [ ]
    get:
      summary: 菜单树列表
      operationId: getMenuTrees
      tags:
        - menu
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Menu"
          description: "获取成功"
      security:
        - bearer: [ ]
  /menus/{id}:
    get:
      summary: 获取菜单
      operationId: getMenu
      tags:
        - menu
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Menu"
      security:
        - bearer: [ ]
    put:
      summary: 更新菜单
      operationId: updateMenu
      tags:
        - menu
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuUpdate"
      responses:
        "204":
          description: ""
      security:
        - bearer: [ ]
    delete:
      summary: 删除菜单
      operationId: deleteMenu
      tags:
        - menu
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: ""
      security:
        - bearer: [ ]
  /menus/tree:
    get:
      summary: 菜单树
      operationId: getMenuTree
      tags:
        - menu
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tree"
      security:
        - bearer: [ ]
  /menus/compact:
    get:
      summary: 菜单树精简
      operationId: getMenuTreeCompact
      tags:
        - menu
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuCompact"
      security:
        - bearer: [ ]
...